% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check.scalars}
\alias{check.scalars}
\title{Assess whether scalar estimates fall within their intended confidence intervals}
\usage{
check.scalars(ecdfs, expected, level)
}
\description{
Assess whether scalar estimates fall within their intended confidence intervals
}
\keyword{internal}
