% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{confints.outcomes}
\alias{confints.outcomes}
\title{Calculate confidence intervals of model outcome variables}
\usage{
confints.outcomes(data, model, level)
}
\arguments{
\item{data}{data frame containing predictor and outcome variables}

\item{model}{associated model}

\item{level}{confidence level}
}
\value{
data frame of predictors and associated lower CL, median, and upper CL
}
\description{
Calculate confidence intervals of model outcome variables
}
\keyword{internal}
