% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ecdf.outcomes}
\alias{ecdf.outcomes}
\title{Return empirical CDF for model outcome variables}
\usage{
ecdf.outcomes(data, model)
}
\arguments{
\item{data}{data.frame containing predictor variables and outcome variables}

\item{model}{associated model (its formula is used to determine which variables in \code{data} are predictors and which are outcomes)}
}
\value{
list of ECDFs
}
\description{
Return empirical CDF for model outcome variables
}
\keyword{internal}
