% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{eval.model}
\alias{eval.model}
\title{Eval model prediction with given model params at given predictors}
\usage{
eval.model(params, model, predictors, outcome)
}
\arguments{
\item{params}{model params}

\item{model}{model to evaluate}

\item{predictors}{predictor values at which to evaluate}

\item{outcome}{outcome function}
}
\value{
data frame of model evaluated with params at predictors
}
\description{
Eval model prediction with given model params at given predictors
}
\keyword{internal}
