% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{observed.results.scalars}
\alias{observed.results.scalars}
\title{Run simulation study on one set of observations}
\usage{
observed.results.scalars(
  observed,
  truth,
  expected,
  fun.model,
  fun.outcome,
  n.samples,
  level
)
}
\description{
Run simulation study on one set of observations
}
\keyword{internal}
