% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar.R
\name{pspline.confints.scalars}
\alias{pspline.confints.scalars}
\title{Calculates confidence intervals and medians for scalar samples obtained from \code{\link{pspline.sample.scalars}}}
\usage{
pspline.confints.scalars(samples, model, level = 0.95)
}
\arguments{
\item{samples}{data frame of samples as returned by \code{\link{pspline.sample.scalars}}}

\item{level}{confidence level for calculated confidence intervals}
}
\value{
data.frame of confidence intervals and medians
}
\description{
Calculates confidence intervals and medians for scalar samples obtained from \code{\link{pspline.sample.scalars}}
}
\keyword{internal}
