% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{pspline.confints.timeseries}
\alias{pspline.confints.timeseries}
\title{Calculates confidence intervals for results of simulation performed by \code{\link{pspline.sample.timeseries}}}
\usage{
pspline.confints.timeseries(samples, model, level = 0.95)
}
\arguments{
\item{samples}{data frame of samples as returned by \code{\link{pspline.sample.timeseries}}}

\item{level}{confidence level for calculated confidence intervals}
}
\value{
data.frame of confidence intervals
}
\description{
Calculates confidence intervals for results of simulation performed by \code{\link{pspline.sample.timeseries}}
}
\keyword{internal}
