% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{pspline.outbreak.calc.cumcases}
\alias{pspline.outbreak.calc.cumcases}
\title{Calculate cumulative incidence time series from incidence time series}
\usage{
pspline.outbreak.calc.cumcases(time, cases)
}
\arguments{
\item{time}{vector of times}

\item{cases}{vector of corresponding incidences}
}
\value{
vector of corresponding cumulative incidences
}
\description{
Correctly handles accumulating over time intervals different from 1
}
