% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{pspline.outbreak.cases}
\alias{pspline.outbreak.cases}
\title{Calculate cumulative incidence for an outbreak}
\usage{
pspline.outbreak.cases(model, data)
}
\arguments{
\item{model}{model returned by \code{\link[mgcv]{gam}} or \code{\link[mgcv]{gamm}}, with a single parameter (time)}

\item{data}{data frame of predictor values at which the model will be evaluated}
}
\value{
data frame of predictor values with corresponding cumulative incidence estimates in \code{$cumcases}
}
\description{
This is useful as \code{outcome} for \code{\link{pspline.estimate.timeseries}}.
}
