% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak.R
\name{pspline.outbreak.thresholds}
\alias{pspline.outbreak.thresholds}
\title{Calculate outbreak thresholds for an outbreak}
\usage{
pspline.outbreak.thresholds(onset = NA, offset = NA)
}
\arguments{
\item{onset}{onset threshold (as fraction of total outbreak case count)}

\item{offset}{offset threshold (as fraction of total outbreak case count)}
}
\value{
function suitable as outcome estimator parameter of \code{\link{pspline.estimate.scalars}}
}
\description{
The result of calling this is useful as \code{outcomes} for \code{\link{pspline.estimate.scalars}}.
}
