% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sample.params}
\alias{sample.params}
\title{Sample spline parameters}
\usage{
sample.params(model, samples)
}
\arguments{
\item{model}{spline GAM}

\item{samples}{number of samples}
}
\description{
Sample spline parameters
}
\keyword{internal}
