% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{truth.results.scalars}
\alias{truth.results.scalars}
\title{Run simulation study on one truth}
\usage{
truth.results.scalars(
  truth,
  fun.observations,
  n.observations,
  fun.model,
  fun.outcome,
  n.samples,
  level
)
}
\description{
Run simulation study on one truth
}
\keyword{internal}
